requirements.txt
# Зависимости
pytest>=8.4.1
selenium>=4.35.0
webdriver-manager>=4.0
allure-pytest>= 2.15.0
